+++ tests/testthat/test-db-example-analysis.R
@@ -68,7 +68,7 @@ check_mutate <- function(cdm) {
 
 }
 
-check_summarise_dplyr <- function(cdm){
+check_summarise_dplyr <- function(cdm) {
 
   expect_true(all(c("min",  "mean", "max") %in%
                    colnames(cdm$person %>%
@@ -193,11 +193,12 @@ check_row_number <-  function(cdm){
 
 }
 
-dbtype = "snowflake"
 checks <- list()
-for (dbtype in dbToTest) {
-  test_that(glue::glue("{dbtype} - checking common usage"), {
 
+for (dbtype in dbToTest) {
+  if (dbtype == "bigquery") next
+  require(dplyr)
+  test_that(glue::glue("{dbtype} - checking common usage"), {
     con <- get_connection(dbtype)
     cdm_schema <- get_cdm_schema(dbtype)
     write_prefix <- paste0("test", as.integer(Sys.time()), "_")