+++ NEWS.md
@@ -1,8 +1,11 @@
 # CDMConnector 1.1.0
 
-* add `copy_cdm_to`
-* add `generate_concept_cohort_set`
-* refactor cross database tests
+* add `copy_cdm_to`, `copyCdmTo`
+* add `generate_concept_cohort_set`, `generateConceptCohortSet`
+* add more example CDM datasets
+* add `record_cohort_attrition`, `recordCohortAttrition`
+* improve database test coverage
+* update vignettes
 
 Released Aug 11, 2023